pyproject.toml

[project]
name = "strawberry-github-bot"
version = "0.1.0"
description = ""
authors = [{ name = "Patrick Arminio", email = "patrick.arminio@gmail.com" }]
dependencies = ["strawberry-graphql", "uvicorn", "starlette", "httpx", "mypy"]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.isort]
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
force_grid_wrap = 0
line_length = 88
multi_line_output = 3


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "flake8<4.0.0,>=3.8.4",
    "black<21.0,>=20.8b1",
    "isort<6.0.0,>=5.7.0",
    "taskipy<2.0.0,>=1.6.0",
    "pdbpp>=0.10.3",
]

[tool.pdm.scripts]
dev.env = { "GITHUB_TOKEN" = "demo", "SKIP_TOKEN_CHECK" = "true" }
dev.cmd = "python main.py --reload"

[tool.pdm.build]
includes = []

[tool.ruff]
line-length = 88
select = ["ALL"]
ignore = ["D", "ANN101"]
target-version = "py311"
